refactor: change password method removed (updated method instead created)
