+++ user/internal/repository/user/dao/user.go
@@ -1,4 +1,4 @@
-package converter
+package dao
 
 import (
 	"user/internal/model"
