+++ user/internal/repository/user/repository.go
@@ -5,7 +5,7 @@ import (
 	"github.com/jackc/pgx/v4/pgxpool"
 	"user/internal/model"
 	"user/internal/repository"
-	"user/internal/repository/user/converter"
+	"user/internal/repository/user/dao"
 	repoModel "user/internal/repository/user/model"
 )
 
@@ -38,13 +38,13 @@ func (u *userRepo) Get(ctx context.Context, email string) (*model.User, error) {
 		return nil, err
 	}
 
-	return converter.ToUserFromRepo(&user), nil
+	return dao.ToUserFromRepo(&user), nil
 }
 
-func (u *userRepo) ChangePassword(ctx context.Context, email, password string) error {
-	sqlQuery := `UPDATE users SET password = $1 WHERE email = $2`
+func (u *userRepo) Update(ctx context.Context, user *model.User) error {
+	sqlQuery := `UPDATE users SET name = $1, email = $2, password = $3, role = $4, updated_at = $5 WHERE id = $6`
 
-	_, err := u.db.Exec(ctx, sqlQuery, password, email)
+	_, err := u.db.Exec(ctx, sqlQuery, user.Name, user.Email, user.Password, user.Role, user.UpdatedAt, user.ID)
 	if err != nil {
 		return err
 	}
